backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
openai==1.74.0
python-dotenv==1.1.0
pydantic==2.11.3
pydantic_core==2.33.1
requests==2.32.3
numpy==2.2.4
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
threadpoolctl==3.6.0
beautifulsoup4==4.12.3
